Ref #81 Rename some modules and unit tests
